refactor: Move ResultsTable to is own component
